requirements.txt
streamlit==1.35.0
pymongo==4.6.2
python-dotenv==1.0.1
pandas==2.2.2
langchain==0.2.1
llama-cpp-python==0.2.72
langchain-community==0.2.1